refactor (create and move package) : Table Group 도메인 패키지 구조 변경 - 테스트 코드 패키지도 변경된 도메인 패키지구조로 일괄 변경
